# apps/orchestrator/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Outlook AI Orchestrator — API",
    "version": "0.1.0",
    "description": "AI Orchestrator backend for the Outlook AI Orchestrator.\n\n**AI-load minimisation.** `EmailAnalysis.source` tells the caller how an answer was produced:\n`llm` (model called now), `cache` (identical content already analysed), `precomputed`\n(analysed ahead of time by the mailbox-sync worker) or `heuristic` (rules only — a triaged\nnewsletter/notification, or the model was unavailable). `EmailAnalysis.triage.kind` says why.\n\n**Human-in-the-loop.** The AI never sends or deletes anything; every action goes through\n`/actions/propose` → human approval → `/actions/approve`."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Azure AD access token (`AUTH_MODE=aad`) or `ADMIN_API_TOKEN`."
      },
      "devHeaders": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-email",
        "description": "Dev identity (`AUTH_MODE=dev` only, refused in production)."
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "correlationId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "Health": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "degraded",
              "down"
            ]
          },
          "checks": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "enum": [
                    "ok",
                    "degraded",
                    "down"
                  ]
                },
                "detail": {
                  "type": "string"
                }
              },
              "required": [
                "status"
              ],
              "additionalProperties": false
            }
          },
          "version": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "checks",
          "version",
          "timestamp"
        ],
        "additionalProperties": false
      },
      "FeatureFlags": {
        "type": "object",
        "properties": {
          "graphEnabled": {
            "type": "boolean"
          },
          "embeddingsEnabled": {
            "type": "boolean"
          },
          "llmProvider": {
            "type": "string"
          },
          "llmModel": {
            "type": "string"
          },
          "llmFastModel": {
            "type": "string"
          },
          "embeddingModel": {
            "type": "string"
          },
          "authMode": {
            "type": "string",
            "enum": [
              "dev",
              "aad"
            ]
          },
          "precomputeEnabled": {
            "type": "boolean",
            "default": false
          },
          "dailyBriefEnabled": {
            "type": "boolean",
            "default": false
          },
          "organizationName": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "graphEnabled",
          "embeddingsEnabled",
          "llmProvider",
          "llmModel",
          "authMode",
          "version"
        ],
        "additionalProperties": false
      },
      "SystemStatus": {
        "type": "object",
        "properties": {
          "health": {
            "$ref": "#/components/schemas/Health"
          },
          "features": {
            "$ref": "#/components/schemas/FeatureFlags"
          },
          "llmQueue": {
            "type": "object",
            "properties": {
              "pending": {
                "type": "integer"
              },
              "running": {
                "type": "integer"
              },
              "concurrency": {
                "type": "integer"
              },
              "avgLatencyMs": {
                "type": "number"
              },
              "circuitOpen": {
                "type": "boolean"
              }
            },
            "required": [
              "pending",
              "running",
              "concurrency",
              "circuitOpen"
            ],
            "additionalProperties": false
          },
          "cache": {
            "type": "object",
            "properties": {
              "analysisHits": {
                "type": "integer"
              },
              "analysisMisses": {
                "type": "integer"
              },
              "embeddingHits": {
                "type": "integer"
              },
              "embeddingMisses": {
                "type": "integer"
              }
            },
            "required": [
              "analysisHits",
              "analysisMisses",
              "embeddingHits",
              "embeddingMisses"
            ],
            "additionalProperties": false
          },
          "sync": {
            "$ref": "#/components/schemas/MailboxSyncStatus"
          },
          "uptimeSeconds": {
            "type": "number"
          }
        },
        "required": [
          "health",
          "features",
          "llmQueue",
          "cache",
          "uptimeSeconds"
        ],
        "additionalProperties": false
      },
...
      "MailboxSyncStatus": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "state": {
            "type": "string",
            "enum": [
              "idle",
              "syncing",
              "error",
              "disabled"
            ]
          },
          "lastSyncAt": {
            "type": "string"
          },
          "nextSyncAt": {
            "type": "string"
          },
          "indexedEmails": {
            "type": "integer"
          },
          "precomputedAnalyses": {
            "type": "integer"
          },
          "pending": {
            "type": "integer"
          },
          "lastError": {
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "state",
          "indexedEmails",
          "precomputedAnalyses",
          "pending"
        ],
        "additionalProperties": false
      },
...
      "IndexEmailsRequest": {
        "type": "object",
        "properties": {
          "emails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnalyzeEmailRequest/properties/email"
            },
            "minItems": 1,
            "maxItems": 200
          }
        },
        "required": [
          "emails"
        ],
        "additionalProperties": false
      },
      "IndexEmailsResponse": {
        "type": "object",
        "properties": {
          "indexed": {
            "type": "integer"
          },
          "skipped": {
            "type": "integer"
          },
          "mode": {
            "type": "string",
            "enum": [
              "hybrid",
              "lexical"
            ]
          }
        },
        "required": [
          "indexed",
          "skipped",
          "mode"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "devHeaders": []
    }
  ],
  "paths": {
...
    "/index/emails": {
      "post": {
        "tags": [
          "search"
        ],
        "operationId": "indexEmails",
        "summary": "Index emails for retrieval",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexEmailsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexEmailsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (`validation_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials (`unauthorized`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role (`forbidden`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (`rate_limited`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error (`internal_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/system": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "getSystemStatus",
        "summary": "Runtime status (queue, circuit, caches, sync)",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Mailbox whose sync status is included."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatus"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (`validation_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials (`unauthorized`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role (`forbidden`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (`rate_limited`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error (`internal_error`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  }
}